requirements.txt
awesome-slugify
clickclick
connexion==1.1.4
Flask-Cache>=0.13.1
Flask-limiter==0.9.4
Flask-Migrate==2.0.4
Flask-OAuthlib==0.9.4
Flask-Session==0.3.1
Flask-SQLAlchemy==2.2
gevent==1.2.2
psycogreen
psycopg2
python-dateutil==2.6.0
redis==2.10.5
stups-zign
zmon-cli
